Merged `worker.js` due to issues with worker registration
Since the worker could not be registered from within the cursor script (different origin to original page) I have merged the two together
using a blob. I would like to continue to have a `worker.js` file which is fetched and embedded into the blob; however, I will not be doing this right now
(partly due to the fact that it is 3:30am).